+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## unreleased
+
+### pack
+
+* Updated to the latest Webpack 5.
+* Replaced `optimize-css-assets-webpack-plugin` to `css-minimizer-webpack-plugin`.
+* Fixed ICSS syntax in stylesheets.
+* Added `source-map-loader` for debugging into original source of node_modules libraries that contain sourcemaps
+
+### serve
+
+* Fixed to disable overlay for warnings.
+* Fixed to allow CORS on `webpack-dev-server`
+
+### test
+
+* Updated to Jest 27.
+
+
 ## 4.1.7 (February 18, 2022)
 
 ### pack